Move dsl packages
