+++ build-logic/src/main/kotlin/com/ku_stacks/ku_ring/buildlogic/dsl/AndroidGradleDsl.kt
@@ -1,4 +1,4 @@
-package com.ku_stacks.ku_ring.buildlogic.primitive
+package com.ku_stacks.ku_ring.buildlogic.dsl
 
 import com.android.build.gradle.TestedExtension
 import org.gradle.api.Project
