+++ build-logic/src/main/kotlin/com/ku_stacks/ku_ring/buildlogic/primitive/ComposePlugin.kt
@@ -1,5 +1,12 @@
 package com.ku_stacks.ku_ring.buildlogic.primitive
 
+import com.ku_stacks.ku_ring.buildlogic.dsl.android
+import com.ku_stacks.ku_ring.buildlogic.dsl.debugImplementation
+import com.ku_stacks.ku_ring.buildlogic.dsl.implementation
+import com.ku_stacks.ku_ring.buildlogic.dsl.library
+import com.ku_stacks.ku_ring.buildlogic.dsl.libs
+import com.ku_stacks.ku_ring.buildlogic.dsl.testImplementation
+import com.ku_stacks.ku_ring.buildlogic.dsl.version
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.dependencies
