+++ build-logic/src/main/kotlin/com/ku_stacks/ku_ring/buildlogic/primitive/FirebasePlugin.kt
@@ -1,5 +1,10 @@
 package com.ku_stacks.ku_ring.buildlogic.primitive
 
+import com.ku_stacks.ku_ring.buildlogic.dsl.bundle
+import com.ku_stacks.ku_ring.buildlogic.dsl.implementation
+import com.ku_stacks.ku_ring.buildlogic.dsl.implementationPlatform
+import com.ku_stacks.ku_ring.buildlogic.dsl.library
+import com.ku_stacks.ku_ring.buildlogic.dsl.libs
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.dependencies
